Refactor tests to use ES6 imports and reorganize helper functions
Refactor authHelpers to use ES6 exports and simplify function declarations
