+++ tests/lib/navigationHelpers.js
@@ -1,0 +1,7 @@
+export const scrollDown = async (page) => {
+  await page.evaluate(() => window.scrollTo(0, document.body.scrollHeight));
+};
+
+export const scrollUp = async (page) => {
+  await page.evaluate(() => window.scrollTo(0, 0));
+};
