+++ tests/lib/selectors.js
@@ -1,0 +1,3 @@
+export const subscriptionHeading = (page) => {
+  return page.getByRole("heading", { name: "Subscription", exact: true });
+};
